Qualify slot identity by panel version across Calyx
SlotId is panel-local, so any persisted artifact or cross-panel map keyed
on a bare slot ordinal was silently wrong once the panel version changed.
Introduce PanelSlotId = (panel_version, SlotId) in calyx-core and thread it
through the crates that persist or join on slot identity.

Rather than reinterpreting unqualified legacy rows, the affected paths now
reject them with explicit remediation:

- calyx-loom: AgreementEdge and measured_tags key on PanelSlotId; weave()
  takes a panel_version. New CALYX_LOOM_PANEL_SCOPE_REQUIRED and
  CALYX_LOOM_XTERM_SCHEMA_UNSUPPORTED direct operators to rebuild cross
  terms from panel-qualified Base rows.
- calyx-sextant: RRF profiles no longer hardcode slot-ordinal tables.
  lookup() resolves slots against the live panel (modality, shape, axis
  names, active + non-retrieval-only) and fails with
  CALYX_SEXTANT_PROFILE_UNBOUND when nothing matches. Drops the AP-60
  hardcoded temporal-slot constants; WeightedProfile carries panel_version.
- calyx-search/persisted, calyx-anneal/tune, calyx-sextant diskann concat:
  signature and key-type propagation for the above.

Add two calyx-registry modules backing panel-generation correctness:

- panel_generation.rs: durable CAS allocator for globally-unique panel
  generation numbers, with an operation_id ledger for idempotent retries
  (64 CAS attempts, 10k owner cap).
- temporal_panel_catalog.rs: durable registration of retrieval-only
  temporal sidecar panels per exact source-panel generation, building a
  fixed E2_recency/E3_periodic/E4_positional template. These are
  query-time-only and never feed primary recall or dedup.

Delete calyx-sextant index/testutil.rs (synthetic vault fixtures for
tests and benches) per the repo-wide tests-removed doctrine; module
declaration and lib re-exports cleaned up with it.

Verified: cargo fmt --all --check clean, cargo check --workspace clean.

+++ calyx/crates/calyx-lodestar/src/kernel.rs
@@ -64,7 +64,7 @@ impl Default for RecallReport {
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Kernel {
     pub kernel_id: CxId,
-    pub panel_version: u64,
+    pub panel_version: u32,
     pub anchor_kind: Option<String>,
     pub corpus_shard_hash: [u8; 32],
     pub members: Vec<CxId>,
@@ -78,7 +78,7 @@ pub struct Kernel {
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct KernelParams {
-    pub panel_version: u64,
+    pub panel_version: u32,
     pub anchor_kind: Option<String>,
     pub corpus_shard_hash: [u8; 32],
     pub built_at_millis: u64,
